Memperbaiki fitur daya, pembayaran dan pengisian pulsa listrik

+++ app/Http/Controllers/ElectricityTopUpController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\ElectricityTopUp;
+use App\Models\ElectricalPower;
 use App\Models\Location;
 use App\Models\Sale;
 use App\Models\LocationPhoto;
@@ -28,51 +29,56 @@ class ElectricityTopUpController extends Controller
         if(Gate::allows('isElectricity') && Gate::allows('isWorkshopRead')){
             $areas = Area::with('locations')->get();
             $cities = City::with('locations')->get();
-            $sales = Sale::with('location')->get();
             $media_sizes = MediaSize::with('locations')->get();
             $media_categories = MediaCategory::with('locations')->get();
-            $electricity_top_ups = ElectricityTopUp::with('location')->get();
+            $electrical_powers = ElectricalPower::with('electricity_top_ups')->get();
+            $locations = Location::with('electrical_powers')->get();
             return response()-> view ('electricity-top-ups.index', [
-                'locations'=>Location::filter(request('search'))->area()->city()->condition()->category()->sortable()->paginate(30)->withQueryString(),
+                'electricity_top_ups'=>ElectricityTopUp::filter(request('search'))->area()->city()->month()->year()->sortable()->paginate(30)->withQueryString(),
                 'areas'=>Area::all(),
                 'cities'=>City::all(),
                 'title' => 'Daftar Data Pengisian Pulsa Listrik',
-                compact('areas', 'cities', 'media_sizes', 'media_categories', 'electricity_top_ups', 'sales')
+                compact('areas', 'cities', 'media_sizes', 'media_categories', 'electrical_powers', 'locations')
             ]);
         } else {
             abort(403);
         }
     }
 
-    public function showElectricityTopup(String $locationId): View
+    public function showElectricityTopup(String $electricalId): View
     { 
         if(Gate::allows('isElectricity') && Gate::allows('isWorkshopRead')){
-            $dataTopUps = ElectricityTopUp::where('location_id', $locationId)->get();
-            $location = Location::where('id', $locationId)->firstOrFail();
             $areas = Area::with('locations')->get();
             $cities = City::with('locations')->get();
-            $media_categories = MediaCategory::with('locations')->get();
             $media_sizes = MediaSize::with('locations')->get();
+            $media_categories = MediaCategory::with('locations')->get();
+            $electrical_power = ElectricalPower::findOrFail($electricalId);
+            $locations = Location::with('electrical_powers')->get();
+            $topUps = ElectricityTopUp::where('electrical_power_id', $electricalId)->year()->get();
             return view ('electricity-top-ups.show-top-ups', [
-                'top_ups' => $dataTopUps,
-                'location' => $location,
-                'title' => 'Detail Data Pengisian Pulsa Listrik',
-                compact('areas', 'cities', 'media_categories', 'media_sizes')
+                'electrical_power' => $electrical_power,
+                'top_ups' => $topUps,
+                'title' => 'Detail Pengisian Pulsa Listrik',
+                compact('areas', 'cities', 'media_sizes', 'media_categories', 'locations')
             ]);
         } else {
             abort(403);
         }
     }
 
-    public function createElectricityTopUp(String $locationId): View
+    public function createElectricityTopUp(String $electricalId): View
     { 
         if((Gate::allows('isAdmin') && Gate::allows('isElectricity') && Gate::allows('isWorkshopCreate')) || (Gate::allows('isWorkshop') && Gate::allows('isElectricity') && Gate::allows('isWorkshopCreate'))){
-            $location = Location::findOrFail($locationId);
+            $areas = Area::with('locations')->get();
+            $cities = City::with('locations')->get();
+            $media_sizes = MediaSize::with('locations')->get();
+            $media_categories = MediaCategory::with('locations')->get();
+            $electrical_power = ElectricalPower::findOrFail($electricalId);
+            $locations = Location::with('electrical_powers')->get();
             return view ('electricity-top-ups.create', [
-                'location_id' => $locationId,
-                'location' => $location,
-                'location_photo'=>LocationPhoto::where('location_id', $locationId)->where('company_id', $location->company_id)->where('set_default', true)->get()->last(),
-                'title' => 'Menambahkan Data Pengisian Pulsa Listrik'
+                'electrical_power' => $electrical_power,
+                'title' => 'Menambahkan Data Pengisian Pulsa Listrik',
+                compact('areas', 'cities', 'media_sizes', 'media_categories', 'locations')
             ]);
         } else {
             abort(403);
@@ -84,7 +90,24 @@ class ElectricityTopUpController extends Controller
      */
     public function create(): Response
     {
-        //
+        if(Gate::allows('isElectricity') && Gate::allows('isWorkshopRead')){
+            $areas = Area::with('locations')->get();
+            $cities = City::with('locations')->get();
+            $sales = Sale::with('location')->get();
+            $media_sizes = MediaSize::with('locations')->get();
+            $media_categories = MediaCategory::with('locations')->get();
+            $electrical_powers = ElectricalPower::with('electricity_top_ups')->get();
+            $locations = Location::with('electrical_powers')->get();
+            return response()-> view ('electricity-top-ups.select-power', [
+                'electrical_powers'=>ElectricalPower::where('type', 'Prabayar')->filter(request('search'))->area()->city()->get(),
+                'areas'=>Area::all(),
+                'cities'=>City::all(),
+                'title' => 'Daftar Data Pengisian Pulsa Listrik',
+                compact('areas', 'cities', 'media_sizes', 'media_categories', 'sales', 'locations')
+            ]);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -94,20 +117,26 @@ class ElectricityTopUpController extends Controller
     {
         if((Gate::allows('isAdmin') && Gate::allows('isElectricity') && Gate::allows('isWorkshopCreate')) || (Gate::allows('isWorkshop') && Gate::allows('isElectricity') && Gate::allows('isWorkshopCreate'))){
             $validateData = $request->validate([
-                'location_id' => 'required',
+                'electrical_power_id' => 'required',
                 'user_id' => 'required',
                 'topup_date' => 'required',
                 'top_up_nominal' => 'required',
                 'kwh_qty' => 'required',
-                'remaining_kwh_qty' => 'required',
-                'last_kwh_qty' => 'required',
+                'remaining_kwh_qty' => 'nullable',
+                'last_kwh_qty' => 'nullable',
                 'receipt_image' => 'required|image|file|max:1024',
-                'remaining_image' => 'required|image|file|max:1024',
-                'last_image' => 'required|image|file|max:1024'
+                'remaining_image' => 'nullable|image|file|max:1024',
+                'last_image' => 'nullable|image|file|max:1024'
             ]);
+            
             $validateData['receipt_image'] = $request->file('receipt_image')->store('electricity-images');
-            $validateData['remaining_image'] = $request->file('remaining_image')->store('electricity-images');
-            $validateData['last_image'] = $request->file('last_image')->store('electricity-images');
+            if($request->file('remaining_image')){
+                $validateData['remaining_image'] = $request->file('remaining_image')->store('electricity-images');
+            }
+            
+            if($request->file('remaining_image')){
+                $validateData['last_image'] = $request->file('last_image')->store('electricity-images');
+            }
 
             ElectricityTopUp::create($validateData);
 
@@ -123,17 +152,17 @@ class ElectricityTopUpController extends Controller
     public function show(ElectricityTopUp $electricityTopUp): Response
     {
         if(Gate::allows('isElectricity') && Gate::allows('isWorkshopRead')){
-            $location = Location::findOrFail($electricityTopUp->location->id);
-            $cities = City::with('locations')->get();
             $areas = Area::with('locations')->get();
-            $media_categories = MediaCategory::with('locations')->get();
+            $cities = City::with('locations')->get();
             $media_sizes = MediaSize::with('locations')->get();
+            $media_categories = MediaCategory::with('locations')->get();
+            $electrical_power = ElectricalPower::findOrFail($electricityTopUp->electrical_power_id);
+            $locations = Location::with('electrical_powers')->get();
             return response()-> view ('electricity-top-ups.show', [
+                'electrical_power' => $electrical_power,
                 'electricity_top_up' => $electricityTopUp,
-                'location' => $location,
-                'location_photo'=>LocationPhoto::where('location_id', $location->id)->where('company_id', $location->company_id)->where('set_default', true)->get()->last(),
                 'title' => 'Detail Pengisian Pulsa Listrik',
-                compact('location', 'cities', 'areas', 'media_sizes', 'media_categories')
+                compact('areas', 'cities', 'media_sizes', 'media_categories', 'locations')
             ]);
         } else {
             abort(403);
@@ -146,12 +175,17 @@ class ElectricityTopUpController extends Controller
     public function edit(ElectricityTopUp $electricityTopUp): Response
     {
         if((Gate::allows('isAdmin') && Gate::allows('isElectricity') && Gate::allows('isWorkshopEdit')) || (Gate::allows('isWorkshop') && Gate::allows('isElectricity') && Gate::allows('isWorkshopEdit'))){
-            $location = Location::findOrFail($electricityTopUp->location->id);
+            $areas = Area::with('locations')->get();
+            $cities = City::with('locations')->get();
+            $media_sizes = MediaSize::with('locations')->get();
+            $media_categories = MediaCategory::with('locations')->get();
+            $electrical_power = ElectricalPower::findOrFail($electricityTopUp->electrical_power_id);
+            $locations = Location::with('electrical_powers')->get();
             return response()-> view ('electricity-top-ups.edit', [
+                'electrical_power' => $electrical_power,
                 'electricity_top_up' => $electricityTopUp,
-                'location' => $location,
-                'location_photo'=>LocationPhoto::where('location_id', $location->id)->where('company_id', $location->company_id)->where('set_default', true)->get()->last(),
-                'title' => 'Edit Data Pembelian Pulsa Listrik'
+                'title' => 'Menambahkan Data Pengisian Pulsa Listrik',
+                compact('areas', 'cities', 'media_sizes', 'media_categories', 'locations')
             ]);
         } else {
             abort(403);
@@ -169,17 +203,17 @@ class ElectricityTopUpController extends Controller
                 'topup_date' => 'required',
                 'top_up_nominal' => 'required',
                 'kwh_qty' => 'required',
-                'remaining_kwh_qty' => 'required',
-                'last_kwh_qty' => 'required'
+                'remaining_kwh_qty' => 'nullable',
+                'last_kwh_qty' => 'nullable'
             ];
             if($request->file('receipt_image')){
-                $rules['receipt_image'] = 'required|image|file|max:1024';
+                $rules['receipt_image'] = 'nullable|image|file|max:1024';
             }
             if($request->file('remaining_image')){
-                $rules['remaining_image'] = 'required|image|file|max:1024';
+                $rules['remaining_image'] = 'nullable|image|file|max:1024';
             }
             if($request->file('last_image')){
-                $rules['last_image'] = 'required|image|file|max:1024';
+                $rules['last_image'] = 'nullable|image|file|max:1024';
             }
 
             $validateData = $request->validate($rules);
@@ -189,17 +223,21 @@ class ElectricityTopUpController extends Controller
                 $validateData['receipt_image'] = $request->file('receipt_image')->store('electricity-images');
             }
             if($request->file('remaining_image')){
-                Storage::delete($request->oldRemainingImage);
+                if($request->oldRemainingImage){
+                    Storage::delete($request->oldRemainingImage);
+                }
                 $validateData['remaining_image'] = $request->file('remaining_image')->store('electricity-images');
             }
             if($request->file('last_image')){
-                Storage::delete($request->oldLastImage);
+                if($request->oldLastImage){
+                    Storage::delete($request->oldLastImage);
+                }
                 $validateData['last_image'] = $request->file('last_image')->store('electricity-images');
             }
 
             ElectricityTopUp::where('id', $electricityTopUp->id)->update($validateData);
 
-            return redirect('/show-electricity-top-up/'.$electricityTopUp->location->id)->with('success','Data pengisian pulsa listrik berhasil dirubah');
+            return redirect('/workshop/electricity-top-ups')->with('success','Data pengisian pulsa listrik berhasil dirubah');
         } else {
             abort(403);
         }
@@ -221,7 +259,7 @@ class ElectricityTopUpController extends Controller
                 Storage::delete($electricityTopUp->last_image);
             }
             ElectricityTopUp::destroy($electricityTopUp->id);
-            return redirect('/show-electricity-top-up/'.$electricityTopUp->location->id)->with('success', 'Data pengisian pulsa listrik berhasil dihapus');
+            return redirect('/workshop/electricity-top-ups')->with('success', 'Data pengisian pulsa listrik berhasil dihapus');
         } else {
             abort(403);
         }
